Move getTextColor to a single source, align maximum on portfolio and Borrow page, and color the Health Impact HFs. Align left-hand HF slider extreme with current HF, rather than hundredth flooring.
